+++ src/messages/infra/message.controller.ts
@@ -1,15 +1,14 @@
-import { cron } from "@elysiajs/cron";
 import Elysia from "elysia";
 
 import { repositories } from "@/shared/singleton/repositories";
 import { sendDailyReport } from "../application/send-daily-report.usecase";
 import { sendFloorWarningMessageUseCase } from "../application/send-floor-warning-message.usecase";
 
+// Rotas apenas para os testes de integração
 export const MessageController = new Elysia({
   prefix: "/messages",
   tags: ["Messages"],
 })
-  //rotas apenas para o teste de integração
   .post("send-daily-report", async ({ set }) => {
     try {
       await sendDailyReport(
@@ -62,42 +61,4 @@ export const MessageController = new Elysia({
         description: "Send a WhatsApp flood alert to a single recipient",
       },
     }
-  )
-  .use(
-    cron({
-      name: "automated-flood-check",
-      pattern: "0 */15 * * * *", // A cada 15 minutos
-      timezone: "America/Belem",
-      async run() {
-        console.log(
-          `[CRON] Verificando níveis de água em ${new Date().toISOString()}`
-        );
-
-        try {
-          await sendFloorWarningMessageUseCase(
-            repositories.sensorRepository,
-            repositories.userRepository,
-            repositories.placeRepository
-          );
-
-          console.log("[CRON] Verificação de enchentes concluída com sucesso");
-        } catch (error) {
-          console.error(
-            "[CRON] Erro na verificação automática de enchentes:",
-            error
-          );
-        }
-      },
-    })
   );
-// .use(
-//     cron({
-//         name: "cron-health-status",
-//         pattern: "*/5 * * * * *",
-//         timezone: "America/Belem",
-
-//         async run() {
-//             console.log(`[CRON] Verificando status de saúde em ${new Date().toISOString()}`);
-//         },
-//     }),
-// );
